confirm delete working

+++ public/scripts/client.js
@@ -36,26 +36,26 @@ $(document).ready(function() {
         }); //end ajax
         //add delay
         getTask();
-        //changecolor();
     }); //end completeTaskBtn
 
     //start deleteTaskBtn
     $('#outputDiv').on('click', '.delete-task-btn', function() {
         var id = $(this).attr('data');
         console.log("deleteTaskBtn clicked", id);
-        var objectToSend = {
-            id: id
-        };
-        $.ajax({
-            type: 'DELETE',
-            url: '/deleteTask',
-            data: objectToSend,
-            success: function(response) {
-                console.log("delete response: ", response);
-            }, //end success
-        }); //end ajax
-        getTask();
-    }); //end deleteTaskBtn
+        var objectToSend = { id: id };
+    if (confirm("Are you sure you want to delete this task?") === true) {
+           $.ajax({
+               type: 'DELETE',
+               url: '/deleteTask',
+               data: objectToSend,
+               success: function(response) {
+                   console.log("delete response: ", response);
+               }, //end success
+           }); //end ajax
+         } else {
+           return false;}
+           getTask();
+       }); //end deleteTaskBtn
 
     //------------AJAX Calls-------------//
     //postTask (send user input)
@@ -111,49 +111,9 @@ $(document).ready(function() {
         // loop through the todolist and display each task
         var outputText = '';
         for (var i = 0; i < tasks.length; i++) {
-            // console.log("status for loop: ", tasks[i].status);
-            // console.log("tasks[i].id:", tasks[i].id);
-            outputText += "<p><div class='task-display " + tasks[i].status + "'>Task:"+ tasks[i].task + "</br>Status:"+ tasks[i].status +"  "+ "<button class='complete-task-btn'type='button' name='completetask' data=" + tasks[i].id + "> Complete Task </button>" + " " + "<button class='delete-task-btn' type='button' name='deleteTask' data=" + tasks[i].id + '>Delete Task</button></div></p>';
+            outputText += "<h2><div class='"+ tasks[i].status + "'>Task:"+ tasks[i].task + "</br>Status:"+ tasks[i].status +"  "+ "<button class='complete-task-btn'type='button' name='completetask' data=" + tasks[i].id + "> Complete Task </button>" + " " + "<button class='delete-task-btn' type='button' name='deleteTask' data=" + tasks[i].id + '>Delete Task</button></div></h2>';
         } // end for
         $('#outputDiv').html(outputText);
     }; //end displayOnDom
 
-    var changeColor = function() {
-        $(".task-display").css("background-color", "green");
-    };
-
 }); // end $
-
-
-
-// outputText += "<div class='task-display " + tasks[i].status + "'><ul>Task: " + tasks[i].task + "</ul><ul>Status: " + tasks[i].status + "</ul></div>";
-// outputText += "<div><ul><button class='complete-task-btn " + tasks[i].status + "'type='button' name='completetask' data=" + tasks[i].id + "> Complete Task </button>" + " " + "<button class='delete-task-btn' type='button' name='deleteTask' data=" + tasks[i].id + ">Delete Task</button></ul></div>";
-
-// //displayTask
-// var displayTask = function(tasks) {
-//     console.log('in displayTask:', tasks);
-//     // loop through the todolist and display each task
-//     var outputText = '';
-//     for (var i = 0; i < tasks.length; i++) {
-//         console.log("status for loop: ", tasks[i].status);
-//         console.log("tasks[i].id:", tasks[i].id);
-//         if (task[i].status === "complete") {
-//
-//
-//         } else {
-//
-//         }
-//
-//     } // end for
-//     $('#outputDiv').html(outputText);
-// }; //end displayOnDom
-
-// if (confirm("do you want to delete this task?")==="true") {
-//   } else { getTask();}
-
-//confirm("do you want to delete this task?");
-//on complete-task-btn click i want to change the color of the output to Green
-
-
-// (if status = complete then color should be green)
-// (if status = work in progress then color should be red)
